fix: print default arg/flag if equal to 0
When the default value for an arg/flag was set to `0` then the `--help`
was not displaying it, as if the default was not set.

There is probably a typing bug in oclif/config or oclif/parser
(depending on what's the desired outcome). The default values typing
is `string`, however the actual values can be set to numbers.